0123Material/requirements.txt
torch
numpy
tyro
diffusers==0.20.2 # old: 0.19.3
dearpygui
einops
accelerate
gradio
imageio
imageio-ffmpeg
lpips
matplotlib
packaging
Pillow
pygltflib
rembg[gpu,cli]
rich
safetensors
scikit-image
scikit-learn
scipy
tqdm
transformers
trimesh
kiui >= 0.2.3
xatlas
roma
plyfile

# zero123plus
pytorch-lightning==2.1.2 # old: 1.9.5
webdataset
omegaconf
megfile